style(PackageCurationProvider): Move this class into its own file
